+++ src/app/core/common-components/entity-form/dynamic-form-validators/dynamic-validators.service.ts
@@ -18,9 +18,12 @@ import {
   descriptionForDateValidator,
   maxDateValidator,
   minDateValidator,
-  parseToDate,
 } from "./date-validators";
-import { calculateAge } from "../../../../utils/utils";
+import {
+  descriptionForAgeValidator,
+  maxAgeValidator,
+  minAgeValidator,
+} from "./age-validators";
 
 /**
  * creates a pattern validator that also carries a predefined
@@ -53,80 +56,6 @@ export function patternWithMessage(
     }
     return errors;
   };
-}
-
-/**
- * Creates an age range validator (min or max age)
- */
-function createAgeComparisonValidator(
-  ageParamValue: unknown,
-  comparator: (age: number, param: number) => boolean,
-  errorKey: string,
-  paramName: "minAge" | "maxAge",
-  rangeConfig?: { minAge?: unknown; maxAge?: unknown },
-): ValidatorFn {
-  const paramNum = Number(ageParamValue);
-  if (!Number.isFinite(paramNum)) {
-    return () => null;
-  }
-
-  const configuredMinAge = Number(rangeConfig?.minAge);
-  const configuredMaxAge = Number(rangeConfig?.maxAge);
-  const hasMin = Number.isFinite(configuredMinAge);
-  const hasMax = Number.isFinite(configuredMaxAge);
-
-  return (control: AbstractControl) => {
-    const controlDate = parseToDate(control.value);
-    if (!controlDate) {
-      return null;
-    }
-
-    const age = calculateAge(controlDate);
-    if (comparator(age, paramNum)) {
-      return null;
-    }
-
-    return {
-      [errorKey]: {
-        [paramName]: paramNum,
-        currentAge: age,
-        ...(hasMin ? { minAge: configuredMinAge } : {}),
-        ...(hasMax ? { maxAge: configuredMaxAge } : {}),
-      },
-    };
-  };
-}
-
-/**
- * Validator for minimum age
- */
-function minAgeValidator(
-  minAgeValue: unknown,
-  maxAgeValue?: unknown,
-): ValidatorFn {
-  return createAgeComparisonValidator(
-    minAgeValue,
-    (age, param) => age >= param,
-    "minAge",
-    "minAge",
-    { minAge: minAgeValue, maxAge: maxAgeValue },
-  );
-}
-
-/**
- * Validator for maximum age
- */
-function maxAgeValidator(
-  maxAgeValue: unknown,
-  minAgeValue?: unknown,
-): ValidatorFn {
-  return createAgeComparisonValidator(
-    maxAgeValue,
-    (age, param) => age <= param,
-    "maxAge",
-    "maxAge",
-    { minAge: minAgeValue, maxAge: maxAgeValue },
-  );
 }
 
 @Injectable({
@@ -307,21 +236,9 @@ export class DynamicValidatorsService {
       case "maxDate":
         return descriptionForDateValidator("maxDate", validationValue);
       case "minAge":
-        if (
-          Number.isFinite(validationValue?.minAge) &&
-          Number.isFinite(validationValue?.maxAge)
-        ) {
-          return $localize`Age must be between ${validationValue.minAge} and ${validationValue.maxAge} years`;
-        }
-        return $localize`Age must be at least ${validationValue.minAge} years`;
+        return descriptionForAgeValidator("minAge", validationValue);
       case "maxAge":
-        if (
-          Number.isFinite(validationValue?.minAge) &&
-          Number.isFinite(validationValue?.maxAge)
-        ) {
-          return $localize`Age must be between ${validationValue.minAge} and ${validationValue.maxAge} years`;
-        }
-        return $localize`Age must be at most ${validationValue.maxAge} years`;
+        return descriptionForAgeValidator("maxAge", validationValue);
       case "pattern":
         if (validationValue.message) {
           return validationValue.message;
